amzn1.ask.skill/lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
pymongo==3.10.1
dnspython==1.16.0
python-dotenv==0.15.0
Unidecode==1.1.2